feat(document): add `$createModifiedPathsSnapshot()`, `$restoreModifiedPathsSnapshot()`, `$clearModifiedPaths()`
Fix #14268
