+++ lib/stateMachine.js
@@ -41,6 +41,7 @@ StateMachine.ctor = function() {
   };
 
   ctor.prototype = new StateMachine();
+  ctor.prototype.constructor = ctor;
 
   ctor.prototype.stateNames = states;
 
@@ -208,4 +209,31 @@ StateMachine.prototype.forEach = function forEach() {
 StateMachine.prototype.map = function map() {
   this.map = this._iter('map');
   return this.map.apply(this, arguments);
+};
+
+/**
+ * Returns a copy of this state machine
+ *
+ * The function profile can look like:
+ * this.forEach(state1, fn);         // iterates over all paths in state1
+ * this.forEach(state1, state2, fn); // iterates over all paths in state1 or state2
+ * this.forEach(fn);                 // iterates over all paths in all states
+ *
+ * @param {String} [state]
+ * @param {String} [state]
+ * @param {Function} callback
+ * @return {Array}
+ * @api private
+ */
+
+StateMachine.prototype.clone = function clone() {
+  const result = new this.constructor();
+  result.paths = { ...this.paths };
+  for (const state of this.stateNames) {
+    if (!(state in this.states)) {
+      continue;
+    }
+    result.states[state] = this.states[state] == null ? this.states[state] : { ...this.states[state] };
+  }
+  return result;
 };
